pull through single tree - nearly there
